third_party/blink/web_tests/http/tests/websocket/multiple-connections-expected.txt

Test that WebSockets are not subject to the HTTP connection limit.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS socketsOpened is 50
PASS successfullyParsed is true

TEST COMPLETE
